BUG#28221731 SHELL SHOULD NOT PROMPT TO SAVE THE PASSWORD IF IT WAS EXPLICITLY SPECIFIED
Shell prompts if password should be saved even if it's explicitly specified by
the user. This creates potential backward-compatibility issues, as existing
scripts may unexpectedly hang on prompts.

This patch changes this behaviour, passwords are going to be saved only if they
are provided by the user at the password prompt.

Change-Id: I74b450d485e2b8d0d6476111a8e874d31a0d89dd