BUG#28219398 INSTANCE ADDED TO CLUSTER STUCK IN RECOVERY PHASE DUE TO WRONG CHANGE MASTER CMD
Occasionally, when adding an instance to an existing cluster, the instance
gets stuck in the distributed-recovery phase resulting in an immutable
reported status of 'RECOVERING'.
This issue happens every-time the generated password for the replication-user
contains the following sequence of characters: '%s'.

The issue is caused by an internal mechanism to replace the occurrences of
'%s' by '?' in any executed query. This leads to the 'CHANGE MASTER' command
issued to initiate the distributed recovery to have a different value for the
password field of the replication-user, causing a connection error.

This patch fixed the issue by removing the internal mechanism to replace the
occurrences of '%' by '?' in the executed queries and by updating all internal
queries that use '%s' as the wildcard for pattern substitution to use '?'.

Added new unit-test for regression.

Change-Id: I1c40882dbf2e7729918e745438955118e1b8b9a8
(cherry picked from commit 7273a33fdd603ca3f46b55aae6a8aed8dfb882a2)