WL#12006 AdminAPI: Convert mysqlprovision check command to C++
This patch includes the conversion of the mysqlprovision check command to C++,
removing the dependency from this internal tool for the
dba.checkInstanceConfiguration(), dba.configureLocalInstance() and
dba.configureInstance() operations.

A new configuration library was added to provide some required features for
this task: mysqlshdk/libs/config/*. Apart from providing features to
read/write option files, it allows configurations to be changed in a more
robust way by allowing all configuration changes to be applied at once at the
end (instead of continuously along the operation) thus leaving the server in
a consistent state if some error occurred before that step (not changing its
settings).

This patch also includes fixes for the following bugs: BUG#27702439,
BUG#27305806, BUG#28733883, BUG#28768627 and BUG#28737777.

Change-Id: Iffc6b7b01ddadab5df28e305d8ff52f7ecf942b6