BUG#27411526 SQL MODE ENTERS MULTILINE MODE WHEN IT SHOULDN'T
Shell running in SQL mode erroneously entered multi-line mode when an unknown
command was executed or a line containing multiple consecutive delimiters was
provided.

Added check for unknown commands, any line starting with \ character is assumed
to contain a shell command. If an unknown command in encountered, an error is
printed to stderr.

The SQL statement parser has been fixed, in some circumstances it was skipping
a character.

Added regression tests.

Change-Id: Ie7a3079639417b90a06c6fa71fc91c6114586f5b