WL#10775 Screen paging
This patch introduces the support for a pager, an external command which
receives text output of certain MySQL Shell commands and results of
statement execution.

Pager can be set using the shell.option['pager'] option and it defaults
to an empty string, which means pager is not set. The initial value of
this option can be set using (in the order of evaluation):
 - PAGER environment variable,
 - persisted value of shell.option['pager'] option,
 - --pager command line option.

At runtime, this option can be changed using two new commands:
 - \pager [command] (\P [command]) - sets the current pager, called with
   no arguments, restores the initial value of shell.option['pager']
   option.
 - \nopager - disables the pager, same as setting the
   shell.option['pager'] option to an empty string.

If set, pager will receive:
 - text output of \help shell command,
 - results of SQL statements executed in SQL mode.

Two new methods allow to use pager in JavaScript and Python scripting
modes:
 - shell.enablePager() (shell.enable_pager() in Python) - enables usage
   of pager in the current scripting mode. All text output (except for
   prompts and user interaction) is going to be forwarded to the pager.
   This behaviour is in effect until shell.disablePager() is called or
   current scripting mode is changed.
 - shell.disablePager() (shell.disable_pager() in Python) - disables
   pager for the current scripting mode.
These methods use the current value of shell.option['pager'] option,
calling them does not change that value. If shell.option['pager'] option
is changed after shell.enablePager() has been called, new pager is going
to be used for all subsequent text output.

Pager is used only if shell is running in the interactive mode.

If pager is set to an invalid command no error message is issued till
pager is used.

Change-Id: I1d00fb5408ea8cfce06b3edeac51553d9149803f