BUG#27381738 NOT ALL WINDOWS NAMED PIPE CONNECTION METHODS WORKS
Named pipe connections on Windows have been systematized, documentation
has been updated to clearly explain how to specify a connection to named
pipe.

If user is using a URI to connect to named pipe, the same rules as for
other operating systems apply (host and port are not allowed), but on
Windows, the named pipe needs to be prepended with "\\.\", i.e.:
 - \\.\named.pipe
 - (\\.\named.pipe)
 - \\.\d%3A%5Csome%5Cpath%5Cnamed.pipe
 - (\\.\d:\some\path\named.pipe)
 - \\.\d%3A%2Fsome%2Fpath%2Fnamed.pipe
 - (\\.\d:/some/path/named.pipe)

Additionally, if user specifies the host to be equal to ".", either via
command line option (--host=.), via URI (user@.) or via connection options
({"host": "."}), shell will initiate the connection to a named pipe. It will
use the default named pipe (MySQL) or the one specified by the user via
"socket" option.

Change-Id: I053724379ffd4e831158ab58d586b3eb8e0025fd