BUG#27363459 SHELL COMMAND OPTIONS ARE INCONSISTENT
The short command options for session type: -ma, -mc, -mx are out
of standard: short options accept parameters either in the format
of -m<param> or -m <param>. In addition, -m was never meant to be
the option name.

To solve this problem they are being renamed as short aliases of
the long version: --mysql (--mc) and --mysqlx (--mx).

The original options have been deprecated.

This is applicable to the \connect command too.

Change-Id: If0d3fe264d28c6d66848bcf5c733701487c4c839