BUG#27480887 @DEBUG3 MODE DISPLAYS NON-JSON OUTPUT WHEN USING --JSON=RAW
The implementation of the logger has been updated, if shell is launched
with a --log-level option which causes the log to be written to stderr,
it will respect the current value of shell.options.outputFormat option.
If that option is set to a JSON-related value, the log output written to
stderr will be JSON-formatted.

Change-Id: I5c053db9fdaac6e4f134f2003ac02e1e961356a3