Fixed test failures in the following AdminAPI auto tests:
- wl10587_ipwhitelist_should_apply_to_repl_user.js
- wl11344_interactive.js

The internal CHANGE MASTER command uses the system hostname value so if
the target platform network settings are not mapping the hostname to the
IP, the command would fail since a replication user wasn't created for
the hostname.

This patch adds the missing hostname value to the ipWhitelist values
whenever used.

Change-Id: Idf6e347f6c3f96313ee332376e54808c265353a3