Fixed test failures in the following AdminAPI auto tests:
- wl10587_ipwhitelist_should_apply_to_repl_user.js
- wl11344_interactive.js

Depending on the platform settings the hostname value may be mapped to
the IP address instead of the real hostname value, leading to failures
in assertions.
In top of that, group_replication_ip_whitelist does not support
hostnames in MySQL 5.7 causing failures in such platforms.

This patch fixes the issues by making use of the MySQL server
configuration setting "report_host", and setting it to the value of
the host IP address.

Change-Id: I3043457df02c841b37b3ec67580c65cd9efcbf36