BUG#28200499 MYSQLSH GENERATES INVALID JSON WHEN QUORUM LOST
The result of calling dba.get_cluster().status() when quorum was lost
could not be converted to a JSON object, because the string
representation of resultant object obtained via str() function
contained escape sequences.

The reported issue is not limited to the Cluster.status() method, but
affects all arrays and dictionaries returned by the Shell API in Python
mode.

The internal representation of arrays and dictionaries has been fixed,
now calling Python built-in function str() on such object will produce
a human-readable string, while calling repr() will result in a string
where special characters are escaped.

Change-Id: Iedf22e981762b62816069c5662ca6d51a58c5953