BUG#28546510 DEPRECATION MESSAGE SHOWN AS ERROR INSTEAD OF WARNING
This problem existed because of a duplicate logic for printing.

Original logic in Shell_core only handled normal text and errors,
when using --json output that is reflected by an "error" field on
the produced output. This deprecation message was being printed as
an error.

A new printing logic has been put in place which supports info,
warnings, errors, notes. However the new engine (IConsole) was
not being used in all the shell.

This patch is a refactoring to remove the old printing engine and
use IConsole on the entire shell, the deprecation messages will be
printed as warnings.

Change-Id: Iebf732e225f8cd1174358cdf4911a99822e0c6ad