WL#12049 AdminAPI: option to shutdown server when dropping out of the cluster
In 8.0.12 and 5.7.24, Group Replication introduced an option to allow
defining the behaviour of a cluster member whenever it drops the cluster
unintentionally, i.e. not manually removed by the user (DBA). The
behaviour defined can be either automatically shutting itself down or
switching itself to super-read-only (current behaviour). The option is:
group_replication_exit_state_action.

In order to support defining such option, the AdminAPI was extended by
introducing a new optional parameter, named 'exitStateAction', in the
following functions:

- dba.createCluster()
- Cluster.addInstance()

This patch implements the addition of this new optional setting.

New tests added to cover the changes and new functionalities added.

In addition, some code clean-up and refactoring was done. Particularly
with the options dictionary handling of those 2 functions.

Change-Id: Ib9cc0df2a43eb5224c4d4eaaeab74ae109aca73b