BUG#28207565 DBA.REBOOTCLUSTERFROMCOMPLETEOUTAGE DOES NOT USE DEFAULT CLUSTER
On the default interactive mode, whenever using the function
dba.rebootClusterFromCompleteOutage without any parameter, the function
should assume the cluster to be rebooted is the default one but it
doesn't and fails with an error specifying the cluster name '' does not
exist.

The issue is caused by the handling of the interactive layer on which if
the user wants to rejoin the instances detected back to the cluster, or
remove from the metadata, a dictionary of options is added to the
function
callback causing on a failure on the detection of the default cluster
which is based on the call of the function without any parameters.

This patch fixes the issue by obtaining the cluster name on the
interactive layer, for the default cluster case, and passing it back to
the function call.

New test added for regression.

Change-Id: Iedeb74bf8ff23c49f10da63d3d90a160d10374c4