WL#11032 AdminAPI: Configure member weight for automatic primary election on failover
In MySQL 8.0.2 and 5.7.20, Group Replication introduces an option to
control the outcome of the primary election algorithm in
single-primary-mode. With this option, the user can influence the
primary member election by providing a member weight value for each
member node. That weight value is used for electing the primary member
insteasd of the member uuid which is the default method used for the
election.

In order to support defining such option, the AdminAPI was extended by
introducing a new optional parameter, named 'memberWeight', in the
following functions:

- dba.createCluster()
- Cluster.addInstance()

This patch implements the addition of this new optional setting.

New tests added to cover the changes and new functionalities added.

In addition, some code clean-up done and the Option_unpacker class was
extended with a new method 'get_optional_exact()' which allows getting
an optional value without doing any type conversion.

Change-Id: Icc9667462a7903a632d1e49fd31d4fbf7e8c3e3b