Unit Test Fixes:
Updated js_adminapi/wl11344_interactive.js test to use different ports
to set the localAddress given that from time to time the ones used were
clashing in ports in use. Since the ones in used are lower values, this
patch changes the ports to use higher values and avoid the random
clashes.

Change-Id: I6d7ffd0abe1f9f20d2e9e9bb3be4daa71d21ad4d