BUG#28544628 UNABLE TO AUTOMATICALLY FETCH STORED PASSWORD FOR A CLASSIC SESSION WITH DEFAULT
If a classic connection was requested to a server, and neither port nor
socket was specified (defaults were used), shell did save the password,
but subsequent connection would still prompt for a password, even
though it should be already available in the secret store.

The connection options provided by the user were used to query the
secret storage, but a different set of options (with default values
filled in) was used to save the password. This caused subsequent
connection attempts to still require a password, as the server URI
provided by the user did not match the one in secret storage.

Issue was fixed, now connection options provided by the user are used
to both query the secret storage and save the password.

Change-Id: I84393a5fdd82885b2aeaad11cb605319c414e71a