BUG#28199513 SHELL HISTORY DELETE OFF SET HANDLING COUNTER-INTUITIVE
To amend problems mentioned in the bug couple things were done:
- when history is empty after delete it will be cleared
- when delete fails because of wrong range error message was augmented with
  valid range as well as correct form is given ("entry" or "range" depending on
  command arguments).
- when deleting from history tail, current offset is updated so there is no
  artificial gap

Also the delete operation was refactored to be able to accept range in format
"-x" which results in deletion of x last entries.

Also as part of this patch, check table command check was fixed in upgrade
checker to run in all the server versions and skip sys schema.

Change-Id: Ieb4bbec49ae24a498ec6aea91f095e2608b1287e