Upgrading version to 8.0.14
The help tests to the util module had to be updated as they had
expectations using a hardcoded shell version.

An upgrade checker test had to be updated as it expected to be
executed with a server of the same version of the shell, now it is
skipped if that's not the case.

Change-Id: Idca9d29aaecf578da038aa8aa5a49f192f55effe