Fixed AdminAPI tests failures:
- The execution of Dba_common_test.validate_member_weight_supported hangs when
running the full Dba_common_test tests due to a deployed but not used sandbox on
the previous test Dba_common_tests.validate_exit_state_action_supported.

If testutil.deploySandbox() is used two consecutive times to deploy an
instance on same port, the 2nd execution hangs on an infinite loop
at wait_sandbox_dead() waiting for the port to be free.

Change-Id: I1fa36daddb23978e4934b52b142ffe9b639630b8