BUG#28236922 : ADMINAPI WILL NOT RECOGNIZE PRIVILEGES IF ASSIGNED WITH A ROLE
The AdminAPI operation dba.checkInstanceConfiguration() does not recognize
privileges when there are associated to a user through a role. In such case,
a missing privileges error is incorrectly be issued despite the user possessing
all the required privileges.

This patches fixes this issue by adding support for roles when checking user
privileges for the AdminAPI. The check for role privileges is only performed
for 8.0 MySQL servers (prior versions do not support roles). In order the obtain
the active roles information (all combinations and dependencies), the SELECT
privilege is required on the mysql.role_edges and mysql.default_role tables.

Note: The SELECT privilege on the mysql schema was already a requirement for
the InnoDB Cluster administration user, therefore no new privilege requirement
was introduced by this patch.

Tests were added for regression.

Change-Id: Id375a5b934d7c7c4ec66dcdc71c73277f3da646d