BUG#28701263 DEFAULT VALUE OF EXITSTATEACTION TOO DRASTIC
The AdminAPI allows setting the value of group_replication_exit_state_action
in the functions dba.createCluster() and <Cluster>.addInstance() by using the
option 'exitStateAction'.

The default value of this option is not controlled by the AdminAPI and set by
Group Replication instead to the value of ABORT_SERVER.
However, this is too drastic and the AdminAPI should take control of it and set
the default to be READ_ONLY.

This patch fixes the issue by setting the value of
group_replication_exit_state_action to READ_ONLY when exitStateAction is
not used in dba.createCluster() and <Cluster>.addInstance().

Added regression tests and updated previous tests which deployed
instances using the parameter
loose_group_replication_exit_state_action=READ_ONLY in order to avoid
the instances shutting down when dropped out of clusters due to GR's
default value of group_replication_exit_state_action.

Change-Id: Iadb0ba6d75b4958d0819cdde6fdfc9e755b8bd8a