WL#12052 AdminAPI: single/multi primary mode change and primary election
Currently, it's not possible to switch a running cluster from single-primary
mode to multi-primary and vice-versa. Nor to trigger the election of a new
primary in single-primary mode.

In 8.0.13, Group Replication introduced the ability to change from
multi-primary to single-primary (and vice-versa) without any downtime and
ensuring the coordination of members, super-read-only handling, and any other
necessary steps are done transparently. On top of that, the ability to force
the election of a new primary member (in single-primary mode) was also
introduced.

In order to support defining such wide-group configuration changes, the
AdminAPI was extended by with three new commands:

  - <Cluster.>switchToSinglePrimaryMode([instanceDef])
  - <Cluster.>switchToMultiPrimaryMode()
  - <Cluster.>setPrimaryInstance(instanceDef)

This patch implements those three new functions, ensuring every InnoDB cluster
related requirements and checks for the new functions:

  - Admin user permissions check
  - Cluster status check
  - Supported version check
  - Metadata updates
  - auto_increment_* updates

New tests added to cover the changes and new functionalities added.

Change-Id: I69a9691bc2de8f2ed9780db7b2c2e74601b52001