BUG#27677227 SILENT ASSUMPTION WHEN CONFIGURING CLUSTER WITH X PROTOCOL DISABLED
When a cluster is created on a server that doesn't have the X Plugin
enabled, a silent assumption is made about the X Protocol port value.

The port value is calculated based on a default assumption that the X
Protocol port is equal to value of the classic protocol * 10. Such value
in stored in the metadata schema under the addresses column of the
instances table, having the value of 'mysqlX'.

Such assumption cannot be done and the address information related to
the X Protocol should only be stored under 'mysqlX' if the protocol is
enabled.

This patch fixes the issue by making sure that the value of 'mysqlX' is
only stored for instances on which the X Protocol is enabled. Otherwise
no value is stored.

Added unit-tests for regression.

Change-Id: I4081eaf2f83dbe83682663b13f8d3777df9d5d0b
(cherry picked from commit e361a44c5197ac2c3767a2cbaa5ff62510766517)