Set read/write timeout during connection
Sets read/write timeout to the same value as the connection timeout for
the duration of the initial connection, in order to avoid freezes in
case the user connects to the wrong network port and an unexpected
service is listening on it.

Infinite timeout is restored after connection is opened, because queries
can take an unknown amount of time to finish.

Change-Id: I116f6587b00ae7d7b7bd39f021d51376da57c26b