WL#10860 DevAPI : Prepared statement support
Automatic use of prepared statements for SQL and CRUD statements
through the X protocol is introduced by this patch as follows:

1) Any statement will execute normally the first time, when executed
   a second time (without changes on the statement), it will be
   prepared and then executed; additional executions will use the
   prepared statement.

2) If the statement changes the previous logic will restart.

3) Any call to a chained function causes the statement to change except
   for calls to limit(), offset(), bind() or execute() (as applicable).
   These means, those functions should be reusable even on a prepared
   statement, reusing it.

4) If the target server does not support prepared statements, it should
   be detected the first time and the prepared statement logic should no
   longer be used with that session.

To achieve #3 a refactoring was required on the way the chained functions
are enabled/disabled. Now when the dynamic behavior is registered for a
function, it is required to indicate:

a) Functions that should be enabled after calling a function, default none.
b) Functions that should be disabled after calling a function, default none.
c) If the function can be called many times, default false.

This patch also removes left over files for collection index creation which
was converted to a single function (not chainable) on the Collection class a
couple of releases ago.

Change-Id: Iadfd034abcb068789155b77820eb89458691efad