WL#12067 AdminAPI: Define failover consistency
In 8.0.14, Group Replication introduced an option to specify the
failover guarantees (eventual or read_your_writes) when a primary
failover happens in single-primary mode).

In order to support defining such option, the AdminAPI was extended by
introducing a new optional parameter, named 'failoverConsistency', in
the dba.createCluster() function.

This patch implements the addition of this new optional setting.

New tests added to cover the changes and new functionalities added.

Change-Id: I9cf409d20acd0b14d378fcc77f2ff90b42e49ecf