WL#10644 : Cluster.rescan() redesign
This patch includes the refactor of the <cluster>.rescan() AdminAPI operation
according to the new command class structure. The old interactive layer code
was removed and it is handled directly in the main operation code. The base
behaviour was maintained, however new options 'addInstances' and
'removeInstances' were added, respectively to allow users to specify new
instances to be added to the metadata and obselete ones to be removed from the
metadata when wizards (prompts and confirmation) are disabled.

In addition, the <cluster>.rescan() operation was extended to detect changes
in the Group Replication mode (single-primary or multi-primary), i.e. not
matching the mode registered in the metadata. A new 'updateTopologyMode' option
was added to allow the user to specify if the metadata should be updated
accordingly or not.

A new 'interactive' option was also added to allow users to enable/disable
wizards (prompts and confirmation) for each execution of the operation.

Regression tests were added for the new features.

Change-Id: I0b37922bea5211176177239d4f7a6e9a38ef22c6