WL#12648 Extensible Object Support
This patch introduces an Extensible_object class which will enable dynamic
registration of objects exposed as API componenes as well as registration
of functions in those objects.

Registered functions can be defined either in C++, JavaScript or Python and
are globally executable: either in JavaScript or Python shell modes.

In addition, the framework allows registering help information for:
- Objects
- Functions
- Function parameters

The help information will be available through the standard shell help system.

Change-Id: I8e0613ec1c7441020ab924d9314b3cf694f384b8