WL12134 - importJson - Convert BSON Data Types
This patch improves the import add support for additional BSON Types
represented using the extended JSON in strict mode format as specified
in MongoDB. This includes:

- Extraction of timestamp from ObjectID values.
- Conversion of the following BSON Values: Date, Timestamp, NumberInt,
  NumberLong, NumberDecimal, Regex and BinData.
- Control of how NumberDecimal values are converted.
- Control of how Regex values are converted.

Several options were added to control how the conversion is done, being
convertBsonTypes the master key for all of them.

This was achieved by refactoring the JSON loader to work in a more
structured way and allowing handling of the conversion at load time.

Side Change
-----------
Unified implementation of fmttime: the function was coded twice at

- mysqlshdk/libs/utils/strformat.h/cc
- mysqlshdk/libs/utils/utils_general.h/cc

Removed the implementation at utils_general.
Added optional parameter to allow:
- Indicating whether the time should be generated in LOCAL or GMT time
  Default is LOCAL
- Indicating a custom time, if null function takes the system time
  Default is nullptr

Change-Id: I9aea519e4fd6a349d52ac2b2d9f215fbaaccf791