WL#12050 AdminAPI: Define the timeout for evicting cluster members
In 8.0.13, Group Replication introduced an option to allow defining
the timeout for evicting previously active nodes.  In order to support
defining such option, the AdminAPI was extended by introducing a new
optional parameter, named 'expelTimeout', in the dba.createCluster()
function.

This patch implements the addition of this new optional setting.

New tests added to cover the changes and new functionalities added.

Change-Id: I85b5410264918ad8037149c4deb58a3b6eed38b6