BUG#25867733 CHECKINSTANCECONFIGURATION SUCCESS BUT CREATE CLUSTER FAILING IF PFS DISABLED
Having the performance_schema enabled is one of the requirements to have
an instance ready for InnoDB Cluster usage.
However, the internal validations of the AdminAPI weren't checking if the
target instance has the P_S enabled.

More precisely, dba.checkInstanceConfiguration() and its internal use in
dba.createCluster() weren't checking if the P_S is enabled in the target
instance. As so, one could create a cluster but could not run several
management commands on it due to P_S being disabled. For example, the
.status() command.

This patch adds that validation, ensuring the AdminAPI checks are
complete.

Added regressions tests.
Fixed Doxygen build errors.

Change-Id: Iaa08853c9904e66a92a2b0c13e735daa89b6b051