BUG#28899522 CONNECTION PARSING REQUIRES A HOST TO BE SPECIFIED BUT LET IT BE EMPTY
The shell.connect() method requires that the host is specified when
connecting to a server using dictionary with connection options, however
it did not verify if provided host is a valid value.

This patch adds necessary validations, disallowing empty values.
Following will now result in an error:
shell.connect({'user':'root','host':''})

Change-Id: I71b991c1488d5537c7223196f949b53a7596ac07