UT Fix: Tests randomly failed due to the registration of help topic for         testutils.
When support for Extensible objects was added, part of the tests included
registering help data for the testutils module, this caused some of the
help tests to fail because they were not expecting help data for the
testutil object.

This patch refactors (simplifies) the registration of help data for
Extensible_objects and their registered functions, also Extensible_objects
now disable their associated help entries when destroyed.

This solves the random failures on the help tests as Testutils will
disable it's own help entries when it is destroyed.

Change-Id: Ie3e14a374008afb5eab2636bb7858e7dda2ba727