BUG#24942875 CLUSTER.CHECKINSTANCESTATE() UNAWARE OF EXISTING CLUSTER MEMBERSHIP
Whenever executing <Cluster.>checkInstanceState() with a target instance which
is a member of the current cluster the output indicates that the instance is
fully recoverable.

This is misleading and is caused by a missing validation to ensure the target
instance does not belong to the cluster.

This patch adds that missing validation.

In addition:

- Refactored the <Cluster.>checkInstanceState() to use the new AAPI class
  structure.
- Added missing Exceptions to the function online helper description
- Added missing tests to cover all possible scenarios and errors.

Change-Id: I8ffa1d179427a4cba82bf363fb1400d1f296791c