Fixes for Windows platform
With the addition of a new method Cluster.options() there is an
ambiguity between that method and shell.options property. When
registering a help entry, if parent topic is specified, but was not yet
registered, the new topic is added to an orphans list. Later on, when
parent topic is registered, all matching orphans are added as well.
Since orphans are identified using the parent name, 'options' can be
matched to either the new method or existing property.

This fix ensures that orphans cannot be registered to a topic which
cannot have children topics (i.e. a method).

Change-Id: I80643db7f4f77de69fe306b5caea4a446ab4bfa9