BUG#28812763 AUTO CORRECT AUTO_INCREMENT_INCREMENT FOR CLUSTER SIZE > 7 TO PREVENT COLLISIONS
The current algorithm to determine the auto_increment_% values is:
   IF single-primary-mode
       set auto_inc_inc = 1
       set auto_inc_offs = 2
    ELSE
       set auto_inc_inc = 7
       set auto_inc_offs = 1 + server_id%7
    ENDIF

This logic ensures no auto increment collisions for multi-primary clusters up
to a size of 7. With multi-primary not being recommended to InnoDB Cluster
users and GR also defaulting to settings that work only up to 7 members this
seemed to be a reasonable choice:

 - not too many numbers are wasted with small clusters
 - most clusters are expected to have 7 or less members
 - GR is/was not capable of dynamically adjusting settings across members

However, duplicate auto increment values may arise with clusters that have
more than 7 members. Group replication does support up to 9 members by
default. If a user plans to run a cluster with more than 7 members in
multi-primary mode, then a manual reconfiguration of the auto increment
settings is required.

The following algorithm should be extended to:

   IF single-primary-mode
       set auto_inc_inc = 1
       set auto_inc_offs = 2
    ELSE
       set auto_inc_inc = n
       set auto_inc_offs = 1 + server_id % n
         where n is the size of the GR group if > 7, otherwise n = 7.
    ENDIF

Up to now, the following commands were using the right algorithm:

- <Cluster.>rescan()
- <Cluster.>switchToSinglePrimaryMode()
- <Cluster.>switchToMultiPrimaryMode()

However, <Cluster.>addInstance() and <Cluster.>removeInstance() weren't.

This patch ensures that the auto_increment_% values are updated accordingly and
automatically regardless of the cluster topology and size.

Additionally, several AdminAPI classes were refactored to:

- Improve the code modularity and organization
- Standarize the class c-tor arguments
- Improve const-correctness

Change-Id: Ib237c466cff56c156b8a43a345771e991e4b60c8