BUG#28915716 MEMSET IS NOT A PROPER FUNCTION TO ERASE SENSITIVE DATA
The buffers used to temporarily store the passwords obtained from the
user were not being properly cleaned, in some cases memset was being
used to "cleanup" the buffers and in other cases not even that.

memset may be optimized away (under the as-if rules) if the object
modified by this function is not accessed again for the rest of its
lifetime, letting the password data available in the memory buffer.

This patch replaces memset with functions that are excluded by this
optimization, explicit_bzero in linux and SecureZeroMemory in windows
eliminating the potential risk.

Change-Id: Ia5c401db5d9eaea3f8e2b31aabc922930a08f2c8