BUG#28285389 : hostname should be taken from hostname or report_host sysvars
The MySQL Shell uses the host value of the provided connection parameters as the
target hostname used for AdminAPI operations, namely to register the instance in
the metadata (for dba.creatCluster() and cluster.addInstance() functions).
However, the host used for the connection parameters may not match the hostname
that is used/reported by Group Replication (GR). GR uses the value of the
'report_host' system variable when defined (not NULL), otherwise the value for
'hostname'. Thus, the same host value should be used by the AdminAPI.

This patch fixes this issue by using the value of 'report_host' if defined (not
NULL) otherwise the value of 'hostname' to register the target instance in the
metadata and as default for the 'group_replication_local_address', instead of
using the host value from the instance connection parameters.

Another issue was also detected and fixed related with this. When the
'report_host' variable is set to empty, for example by using "report_host=" in
the option file (my.cnf), GR uses an empty value for the host but the AdminAPI
(e.g., dba.checkInstanceConfiguration(), dba.configureInstance(),
dba.createCluster()) reports the 'hostname' as the value used which is
inconsistent with the reported reported by GR. An error is now issued by the
AdminAPI if an empty value is set for the report_host system variable.

The help description for the localAddress option for dba.createCluster() and
cluster.addInstance() was updated to mention the new default for the host value.

Existing tests were updated and new test added for regression.

Change-Id: I2ece5a4983eff752b06035fcef521600e826a4d3