WL#12020 : AdminAPI: Refactor addInstance()
This patch refactors the <cluster>.addInstance() command according to the
newly defined class structure for the AdminAPI, and it removes the dependency
from mysqlprovision (internal Python tool) to execute this operation (i.e.,
converting the join-replicaset command from mysqlprovision to C++). The main
behavior of the operation is maintained.

Existing tests already cover the performed changes.

In addition, it ensures BUG#28056944 is fixed, including a specific test
for regression.

Change-Id: Ie949dbae5a46f925870faf4712f1c3e5d85a8c9a