BUG#29246110 : cluster creation fails in debian-based platforms
In Debian-based platforms (such as Ubuntu), it is not possible to create a
cluster using the default settings, where the hostname resolves to 127.0.1.1.
An error is issued by the Group Replication plugin when starting because the
IP address 127.0.1.1 (used by default in Debian platforms) is not supported.
This situation should be validated and a proper error issued before
configuring the instance and attempting to start Group Replication.

This patch fixes this issue by adding a proper validation to the instance
check performed before creating a cluster and adding instances to it.

Tests were added for regression.

Change-Id: I67cdda677d2007f9349d7d73f44602724c33ef95