WL#12764 JSON resultFormat
This patch fixes the handling of the resultFormat option when any of
json or json/raw is used but JSON Wrapping (--json) is turned OFF.

When using resultFormat=json or json/raw every row/document in the result
will be printed as an independent document.

The format used for result stats and warnings is now the same as with
any of the other result formats.

This patch does NOT change the format when JSON Wrapping is turned ON
with the --json=[pretty|raw] command line option.

Change-Id: I487694a124c66d296aaa079d0a4134584ad86d92