WL#12881: AdminAPI: Provide information about Server version of cluster members
The AdminAPI, as the control panel of InnoDB cluster, provides commands to
obtain the information or status of a cluster. However, the information of each
cluster member MySQL Server version is not available. Since clusters can be set
up with instances of different versions, this information should be provided.

This WL improves the observability capabilities of the AdminAPI by adding the
information about the MySQL Server version of cluster members in the commands:

- <Cluster.>status()
- <Cluster.>describe()
- <Cluster.>rescan()

The information about each cluster member version is available if:

- The object session belongs to a MySQL Server with a version >= 8.0.11.
- The object session belongs to a MySQL Server with a version < 8.0.11 but the
  option 'queryMembers' is enabled (in <Cluster.>status()).

New tests added to cover the changes and new functionalities added and existing
tests updated.

On top of that, the implementation of <Cluster.>describe() was refactored to
follow AdminAPI's Command_interface, some code clean-up was performed, small
issues fixed and the register of <Cluster.>describe() and <Cluster.>status()
was updated to make use of the new interface 'expose()'.

Change-Id: I32a3843833e5fc33ff9e89122522d4b549e65733