WL#11155 - Add shell command for sql execution (\sql)
This WL extends the \sql command to allow quick SQL statements execution
in non SQL modes (JavaScript and Python).

When a user issues a command in the form of "\sql line" the line part is
treated as complete SQL statement and is immediately executed in SQL mode,
but (unlike simple "\sql" command) Shell remains in the mode it was before
user issued the command.

Change-Id: I3c245ae52301bec4b647a411e32a7993dd748d43