BUG#29349014 : Shell removes client section in configuration file
There was an issue in the ConfigureLocalInstance() command. After
executed, it would remove any section that did not start with mysqld
and its respective contents from the provided configuration file.
This problem was caused because of a wrong configuration in the
interaction between the Config_file_handler and the Config_file
objects.
The Config_file_handler was always setting the optional group_prefix
parameter of the read method on the Config_file to mysqld. As a result
it would only read sections that start with mysqld to the internal
configuration map object, but when the time came to write the changes,
it would assume the other sections were to be deleted since they were
not in the internal configuration map.

This patch fixes this issue and extends the apply() method from
Config_file_handler to test this scenario.

Change-Id: I73b5ae3c97c7aa8f75232ed2855b810cad79d13e