Config_file: enabling case sensitive/insensitive for groups.
The Config_file class was created only having in mind the MySQL
configuration files where the option groups are case insensitive.

With the introduction of the OCI it is required to support case
sensitive option groups.

This patch Updates the Config_file class to support both approcahes
by using a 'case' flag at the constructor which by default is
Case::INSENSITIVE to keep compliant with the current use cases.

Change-Id: I5395c28aca9d42165114d28c2503fdaf181a7e17