WL#12765: AdminAPI: automatic handling of GR protocol version
In 8.0.16, the Group Replication Communication System (GCS) introduces support
for data fragmentation. Such support is available on MySQL Server 8.0.16 by
default and it can be enabled/disabled.

However, the group does not automatically downgrade its protocol to
allow an instance that does not support it to join the group.

On top of that, the group also does not automatically set the new protocol
version after a successful upgrade of its members or when a member that has
no support for data fragmentation wants to join.

This WL solves both issues by making the AdminAPI transparently handle the
protocol version without the need for user intervention.

Unit-tests added to cover the features added.

Change-Id: Id55299b456e810a439967ae3c55282cc95e2b5d1