BUG#29304183 : removeInstance() command fails with LogicError
When an instance with disabled X Plugin was added to InnoDB Cluster,
it will later failed to be be removed from it using removeInstance(),
issuing a LogicError "get_string(7): field is NULL". This is a
regression issue introduced by a bug fix in the previous release.

This patch fixes the issue by properly retrieving the instance
information from the metadata, i.e., considering that mysqlx can be
disabled and there is no data about it in the metadata, in order to
be restored (if needed) when the removeInstance() operation fails.

New test was added for regression.

Change-Id: I0a20fbdcc275aa19aeaf3f449a139a0528d2a247